feat: add source iban option and remove main account settings
In order to add the option to allocate money from multiple sources we remove the main account settings. Before we used this id as a source always, now we take source_iban from the allocation and group allocations on source_iban. This allows us to completely remove the settings in firebase. In order to top up the source_iban we use fixed instead.